requirements.txt
appdirs==1.4.3
dj-database-url==0.5.0
Django==1.11
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-extensions==2.1.0
djangorestframework==3.9.0
gunicorn==19.7.1
numpy==1.12.1
packaging==16.8
psycopg2-binary==2.7.7
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
six==1.10.0
django-cors-headers==2.4.0